+++ base/src/main/java/tech/thdev/base/presenter/AbstractPresenter.kt
@@ -3,7 +3,7 @@ package tech.thdev.base.presenter
 /**
  * Created by tae-hwan on 8/17/16.
  */
-abstract class AbstractPresenter<VIEW : BaseView<*>> : BasePresenter<VIEW> {
+abstract class AbstractPresenter<VIEW : BaseView> : BasePresenter<VIEW> {
 
     protected var view: VIEW? = null
         private set