+++ base/src/main/java/tech/thdev/base/presenter/BasePresenter.kt
@@ -3,7 +3,7 @@ package tech.thdev.base.presenter
 /**
  * Created by tae-hwan on 8/17/16.
  */
-interface BasePresenter<in VIEW: BaseView<*>> {
+interface BasePresenter<in VIEW : BaseView> {
 
     /**
      * View Attach.