+++ base/src/main/java/tech/thdev/base/presenter/BaseView.kt
@@ -3,7 +3,6 @@ package tech.thdev.base.presenter
 /**
  * Created by tae-hwan on 8/17/16.
  */
-interface BaseView<in P> {
+interface BaseView {
 
-    fun onPresenter(presenter: P)
 }