+++ base/src/main/java/tech/thdev/base/util/ActivityUtil.kt
@@ -6,19 +6,32 @@ import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
 import android.support.annotation.IdRes
-import android.support.v4.app.Fragment
 import android.support.v7.app.AppCompatActivity
 
 /**
  * Created by tae-hwan on 8/17/16.
  */
-fun AppCompatActivity.replaceContentFragment(@IdRes frameId: Int, fragment: Fragment) {
-    supportFragmentManager.beginTransaction().add(frameId, fragment)?.commit()
+
+/**
+ * Content v4.fragment replace.
+ */
+fun AppCompatActivity.replaceContentFragment(@IdRes frameId: Int, fragment: android.support.v4.app.Fragment) {
+    supportFragmentManager.beginTransaction().replace(frameId, fragment)?.commit()
 }
 
+/**
+ * Content app.fragment replace.
+ */
 fun Activity.replaceContentFragment(@IdRes frameId: Int, fragment: android.app.Fragment) {
     fragmentManager.beginTransaction().replace(frameId, fragment).commit()
 }
+
+/**
+ * find fragment by IdRes.
+ */
+fun AppCompatActivity.getContentFragment(@IdRes frameId: Int)
+        = supportFragmentManager.findFragmentById(frameId)
+
 
 fun <E> Context?.startServiceClass(cls: Class<E>?) {
     this?.let {