+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Change Log
 
+### Version 1.0.2 (2016-09-05)
+
+- Classification by usage.
+ - Not used Presenter : BaseActivity/BaseFragment
+ - Use Presenter : BasePresenterActivity/BasePresenterFragment
+
 ### Version 1.0.1 (2016-09-02)
 
 - Change ActivityUtil